mysite_env/requirements.txt
Django==2.1.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
Pillow==5.2.0
pytz==2018.5
six==1.11.0
